moviemate/requirements.txt
Flask
pyarrow
scikit-learn
pandas
scikit-surprise